restructure and delete some classes
