+++ src/main/java/fr/inria/corese/core/next/api/query/TupleQuery.java
@@ -1,11 +1,12 @@
 package fr.inria.corese.core.next.api.query;
 
+import fr.inria.corese.core.next.api.Value;
 import fr.inria.corese.core.next.api.result.TupleQueryResult;
 
 public interface TupleQuery extends Query {
     TupleQueryResult evaluate();
 
     // Bindings
-    TupleQuery setBinding(String name, Object value); // or Value/IDatatype
+    TupleQuery setBinding(String name, Value value);
     TupleQuery clearBindings();
 }
