--- src/main/java/fr/inria/corese/core/next/api/query/ast/AskQueryAst.java
@@ -1,4 +1,0 @@
-package fr.inria.corese.core.next.api.query.ast;
-
-public class AskQueryAst {
-}
