--- src/main/java/fr/inria/corese/core/next/api/query/parser/SparqlOptions.java
@@ -1,4 +1,0 @@
-package fr.inria.corese.core.next.api.query.parser;
-
-public class SparqlOptions {
-}
