--- src/main/java/fr/inria/corese/core/next/api/query/parser/SparqlParser.java
@@ -1,8 +1,0 @@
-package fr.inria.corese.core.next.api.query.parser;
-
-public class SparqlParser {
-
-    /**
-     * String to ParseTree
-     */
-}
