+++ src/main/java/fr/inria/corese/core/next/impl/sparql/CoreseQueryArtifactAdapter.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql;
+
+public class CoreseQueryArtifactAdapter {
+}
