+++ src/main/java/fr/inria/corese/core/next/impl/sparql/ast/QueryAst.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.ast;
+
+public class QueryAst {
+}
