+++ src/main/java/fr/inria/corese/core/next/impl/sparql/bridge/CoreseAstQueryBuilder.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.bridge;
+
+public class CoreseAstQueryBuilder {
+}
