+++ src/main/java/fr/inria/corese/core/next/impl/sparql/parser/SparqlAstBuilder.java
@@ -1,9 +1,11 @@
-package fr.inria.corese.core.next.api.query.parser;
+package fr.inria.corese.core.next.impl.sparql.parser;
 
 public class SparqlAstBuilder {
 
     /**
      * ParseTree to AST
      */
+
+
 
 }
