+++ src/main/java/fr/inria/corese/core/next/impl/sparql/parser/SparqlListener.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.parser;
+
+public class SparqlListener {
+}
