+++ src/main/java/fr/inria/corese/core/next/impl/sparql/parser/SparqlParser.java
@@ -1,0 +1,17 @@
+package fr.inria.corese.core.next.impl.sparql.parser;
+
+import fr.inria.corese.core.next.impl.sparql.ast.QueryAst;
+
+public class SparqlParser {
+
+    /**
+     * String to ParseTree
+     */
+
+    /*
+    QueryAst parse(String queryString) {
+        ???;
+    }
+
+     */
+}
