+++ src/main/java/fr/inria/corese/core/next/impl/sparql/result/CoreseBinding.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.result;
+
+public class CoreseBinding {
+}
