+++ src/main/java/fr/inria/corese/core/next/impl/sparql/wrapper/CoreseBindingSet.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.wrapper;
+
+public class CoreseBindingSet {
+}
