+++ src/main/java/fr/inria/corese/core/next/impl/sparql/wrapper/CoreseGraphQueryResult.java
@@ -1,0 +1,4 @@
+package fr.inria.corese.core.next.impl.sparql.wrapper;
+
+public class CoreseGraphQueryResult {
+}
